Make ZK component more robust against ZK errors
For some Zookeeper related errors during update of a ZK component the
operation might still succeed despite an exception.

However, as we don't update the zstat in those cases, we can get into a
situation where we can no longer update the component state due to a
mismatch of the znode version. This will result in `BadVersionError`
exceptions.

To recover from that we'll set the local zstat to None on all Zookeeper
exceptions, except for the `BadVersionError`.

Change-Id: Ic06ca08f00a4660902873a4efddb220c5e8a0a14