Use warning color for canceled build results in queue item progressbar
This adds the missing color-coding for a canceled build in the overall
queue item progressbar. Since this is still using an old PF3 progress
bar, the color differentiate slightly from the PF4 warning color.

However, we might want to update this component to PF4 anyways, so this
is more like an intermediate step.

Change-Id: I83285f288e110b3c40fb2dee804079c41c748cbc