WIP: Update queue item progress bar to PF4
The queue item progress bar is a custom component that is different from
the job progress bars. While the latter one were already converted to
PF4, the queue item progress bar was not.

The main reason for this is that the queue item progress bar is actually
a div that contains a row of multiple progress bars - one for each job.
Since the PF4 progress bars are designed completely different than the
PF3 progress bars, we have to make a few styling adaptions to keep the
old design/functionality.

Additionally, this change updates some icon mappings for the job results
and adds missing ones.

Change-Id: I83285f288e110b3c40fb2dee804079c41c748cbc