Update patch set 1
Patch Set 1: Code-Review-1

I really like this change; I think it's a big improvement for the use case of "watch my change as it progresses through the system".

I think there's a mismatch between what the tooltip says and what the change does.  The tooltip says "Filter for this change" but the actual filter is "Filter for this change and this project in this pipeline".

The project is redundant -- a given change id can't move to a different project, so we can drop that from the filter altogether.

The pipeline effectively makes it "Filter for this queue item", and means that if the change moves from check to gate, the user won't see it.

Having said that, perhaps it's nice to be able to automatically add a bunch of filters, then remove the ones you don't want.

Let's look at some use cases:

1) User wants to watch their change progress through the system.  We should only filter for change id.
2) User wants to see all changes for their project.  We should only filter for project.
3) User wants to see exactly this queue item (maybe to know when a merge is done).  We should filter for change and pipeline.

I feel like #1 is a common use case and I think since it's easy to click the pipeline after clicking the funnel, that filtering only for change id serves both #1 and #3.

Likewise the reverse for #3 -- click the funnel then click to delete the pipeline.

Those don't serve use case #2 though; the only one that does is what you have now, then removing the two extra fields for use case #1.  So it's a little more difficult for what may be the most common case but it serves all 3 easily.

There is precedent for this though:

If you visit this page: https://zuul.opendev.org/t/openstack/config-errors

You can see that on the right of each table row, there is a search-plus icon.  We used that icon to indicate we're adding a field to the set of filters.  We may want to use that instead of the filter icon, or if we feel the filter icon is more appropriate, switch that.

If you click on that, you get a popup with the option to filter by several different options.  We could do the same here and make this icon open a popup that says (matching the format in the error page):

Filter by change: 1234
Filter by project: nova
Filter by item: 1234 in gate

We could change the format ("Filter for this change") but if so, we should probably change the error page to match.  This should be consistent.  We have also talked about how we would like to extend this pattern to the build and buildset result pages, and possibly others.

How to proceed?

1) Decide whether to use a popup like the error page.
   My preference is yes, since it serves all three use cases easily and equally.
   If we decide not to, then I think we should add all the filters like this change does now, but then we should also do the same on the error page, so that all 4 filters are added then let users remove them.
   I think it makes more sense to let users add the filters they need rather than add a bunch and ask them to remove the ones they don't.  So if we decide against the popup, I think we need a good reason.

2) Decide which icon to use.  I think this page and the error page should behave the same and have the same icon.  As mentioned earlier, there's a reason we used search-plus when adding individual filters (this pattern and this icon show up in systems like elasticsearch/logstash).  So if we go with the popup, I think we should use search-plus.  If we add all the filters at once, I think we should use the filter icon.  But I don't feel very strongly about either -- only that we are consistent.

3) If we end up deciding to go with the change as written (add multiple filters at once without a popup) we should adjust the tooltip to match.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1, 815aebe11985b17c820eb272506f3a58a0d10fbe