Don't let long project names break the ProgressStepper layout
If a QueueItem contains a very long project name, the title of the
QueueItem card might break the layout as the width of the QueueItem
card would be wider than the width of the surrounding ProgressStepper.
As a result, the QueueItem card could overlay other QueueItems in
another column.

By setting an overflow-wrap on the QueueItem title, we ensure that the
title is wrapped before it breaks the layout.

Change-Id: I26a034e87f5836c1dd30bf93eadf91ab013e6565