Use uploaded image id when launching nodes
This adds support for launching nodes from zuul-built images (ie, dib images).

This corrects some oversights in earlier work:

Endpoints now have canonical names that combine their name with their
connection names.  An example might be "aws/us-west-2".  Some
references to endpoints used only their names which may not be
sufficiently unique.

Since the original spec was written, we have added a considerable
number of driver-specific options to nodepool image configuration.
Originally it was expected that all characteristics of an image would
be described in the global "image" object, and therefore the canonical
name of a given image uploaded to multiple providers would be
sufficiently unique to identify it.  But if we are to entertain the
idea that an image can be uploaded to multiple providers with, for
example, different aws volume performance settings (iops, etc), then
that is no longer sufficient.

To address that, we now identify unique image configurations across
multiple providers.  So if an image appears in multiple providers with
the same configuration, only one upload will occur.  If it is
configured for 200 iops in one provider and 1000 iops in another
provider, two uploads will occur.

The typical cases where global images are attached to a single section
shared across multiple tenants will result in only a single upload per
endpoint.  Likewise, if a tenant adds their own image to a provider,
only one upload per section should occur.

This adds a test for the complete lifecycle, but the actual AWS image
upload is still stubbed out.  That will be tested explicitly in the
aws driver tests.

Change-Id: I68c30bb59c2bf6c25b2c9c8df99016e2e954cf42