Use change label in filter dropdown on QueueItem
So far the change text in the filter dropdown showed either the
change.id or change.ref which could break the layout if a long SHA was
displayed together with a long project name.

The simplest solution would be to display the short SHA instead.
However, the exact values of change.id and change.ref depend on the type
of change (github pull request, gerrit change, revision, branch, ...).

We already evaluate those different types in the ChangeLink component to
show the correct "short" presentation for each type. By moving this
"change evaluation" part into a dedicated function, we can reuse it in
the filter dropdown and align the presenation in both components.

Change-Id: Icf8ab38046a972cb8ecced923b9aa53c9dc7a69c