Fix errors in branch visualization when filters are applied
To visualize the branches in a queue ("subway map") we build a tree
structure that links items to their items_behind. Usually each item only
contains a single item_behind which results in a tree with a single
branch (like a linked list). If an item contains multiple items_behind,
we create a new branch in the tree for each one, resulting in tree-like
structure. A typical use case for this is a merge conflict in the gate
pipeline.

When filters are applied, we remove all items that don't match the
filter criteria. This could result in items_behind not being resolvable
anymore as they were removed from the head. The current tree
implementation still links to those non-existing items which results in
rendering issues. To prevent this, we only add an item_behind to the
tree if it still exists in the head list.

A similar error happens if the first item in gate failed, but doesn't
match the currently applied filters. In this case, a tree will be
created for an empty head list resulting in a different rendering
issue. To solve this, we filter out empty head lists before creating a
tree.

Change-Id: I302923c196434fb082584d094ae5c4a838cbc707