Show correct change type in filter dropdown
This shows the correct change type (change, branch, revision) in the
filter dropdown for a QueueItem.

This is only a change in visualization, it doesn't change which filters
are applied. E.g. if we select the first entry in the filter dropdown
for an item in a periodic pipeline, the dropdown list shows something
like:

  Filter branch: /refs/heads/master

If we click on this, the filter toolbar (and also the url) will still
filter by change like so:

  Change: /refs/heads/master

The next step would be to reflect this change also in the internal
filter logic and while doing so, we might want to align the filter
(categories) on the builds/buildsets pages with the ones on the new
status page.

Change-Id: Idfac29309ae4ed56fc98642667e6884b82c3e747