WIP: Show correct change type in filter dropdown
This shows the correct change type (change, branch, revision) in the
filter dropdown on a QueueItem.

This is only a change in visualization, it doesn't change which filters
are applied. E.g. if we select the first entry in the filter dropdown
for an item in the periodic pipeline, the dropdown list says:

  Filter branch: /refs/heads/master

However, the filter toolbar will still apply the filter:

  Change: /refs/heads/master

Change-Id: Idfac29309ae4ed56fc98642667e6884b82c3e747