Use a consistent zk timeout in tests
The fake nodepool in tests uses the default zk connection timeout
of 10s, but the rest of the test infrastructure uses 30s.  This
means that we most often see spurious errors on slow test nodes
in the fake nodepool component.  It is usually able to recover,
which is why this hasn't been an issue until recently.  But now
that we look for unexpected tracebacks in logs, this is becoming
problematic since it's failing tests.

To address this, use a 30s timeout everywhere.

Change-Id: I3d3fedb22eaa8c389c7faf179d05164b6896517f