Increase zk connection timeout in tests
We currently have a 30s zk timeout in tests, but the actual production
Zuul default value is 120s.  To further reduce the number of incidents
of spurious test failures on slow nodes, increase the timeout used
in tests to 60s (still more conservative than the Zuul default, but
much less conservative than the kazoo default of 10s).

Change-Id: Ic53d5be2d9cf134e6dc5217f5056ee69c3bd4ee3