Adjust some AWS tests
Two tests in test_aws_driver can now be moved to test_launcher since
we have sufficient infrastructure in that file to run these tests, and
they are tests of general launcher functionality, not AWS-specific:

* test_jobs_executed
* test_launcher_failover

The test test_aws_launcher only tested that we can instantiate the
provider object on a launcher.  This is now sufficiently covered by
other tests so it can be removed.

Change-Id: I452d1dd05963ef2a4411d46fc550e1b416a33e2a