Add aws snapshot image upload test
This adds test coverage of the AWS snapshot image upload method
(and fixes some issues preventing it from working).

One section of the upload method is removed from an exception handler
because it is required to succeed by code that follows.  This just
allows for better error reporting in the unlikely case that section
fails (the result is the same).

Change-Id: Ifc05a34e2c8e4bb71eed5ba39be69d2de476556e