Inherit some attributes from provider configs
Images, labels, and flavors can inherit some settings from their
attached providers (in order to reduce duplication in the config
yaml).  The validator was already constructed to allow this, but
the implementation to actually use the values was missing.  This
adds an implementation that uses the values of any inheritable
attributes from the provider if they are not present in the label,
image, or flavor item.

Change-Id: I4956d545fd8df9f82fe04e10d6a2ec2863651855