node model: adapt to node_properties
The nodes are receiving flags from Nodepool
via the node_properties attribute on the
Node instance. These node_properties need to
end up in the nodepool dictionary produced by Zuul
when generating the inventory.yaml file. By having
the properties in the inventory.yaml file, it is possible
at job runtime or in the post analysis to determine certain
details about the node that is/was used for the job.

Change-Id: I977107332ef3fe4d168c419fbcb9c6cd2607b580