Update more ZooKeeperClient timeouts in tests
A recent change updated some uses of the ZooKeeperClient to use
a consistent, and more generous, timeout, in order to reduce the
number of incidents of racy test failures.

There are several more places where we construct a client object
which could benefit from the same change.  This commit updates the
remaining instances.

Change-Id: Ia7b7fc491779145fdf71449e4caa6c331553a937