Avoid launcher busy-looping due to constant wake
Constantly checking the requests/nodes and setting the wake event if
they haven't finished yet might cause the launcher to busy loop and
starve other threads.

To avoid that we'll keep track of how often we've polled a node or
request and increase the interval for re-checking the node (exponential
backoff).

Change-Id: I87990c5c5731b60b3ede8e5cdd6a3695e964ab0e