Add support for 'min-ready' provider nodes
The available 'min-ready' nodes for a provider will be shared by all
tenants that have a particular provider and label configured.

In case the min-ready count is not satisfied for a tenant a random
provider in that tenant will be picked to spawn the missing nodes.

In order to avoid race conditions when creating min-ready nodes, the
launchers will calculate a score for the canonical label name, with the
launcher that has the highest score being responsible for creating
the missing nodes.

Another option would be share all nodes of an endpoint/connection that
are using the same label configuration (independent of the provider).
For this we'd need to resolve the full label specification (including
for example flavor and image) and store a hash of that as part of the
provider node. However, it seem like this could introduce other issues
e.g. min-ready nodes being invalidated when the config changes.

Change-Id: I14d61719da6d4719cd147edeb6b35d7cae4ee93e