Add support for 'min-ready' provider nodes
The configured 'min-ready' for a label will be shared by all tenants
that have a provider using the same combination of canonical label and
connection.

In case the min-ready count is not satisfied for a tenant a randome
provider in that tenant will be picked to spawn the missing nodes.

Change-Id: I14d61719da6d4719cd147edeb6b35d7cae4ee93e