Add support for 'min-ready' provider nodes
The available 'min-ready' nodes for a provider will be shared by all
tenants that have the specific label configuration and
connection/endpoint configured.

A 'min-ready' nodes is considered usable if its label config hash
is available on any of the request tenant's providers and has the same
connection/endpoint.

In case the min-ready count is not satisfied for a tenant a random
provider in that tenant will be picked to spawn the missing nodes.

In order to avoid race conditions when creating min-ready nodes, the
launchers will calculate a score for the canonical label name, with the
launcher that has the highest score being responsible for creating
the missing nodes.

TODO:
* add more tests that exercise especially the config hash path
* decide on how we want to deal with existing min-ready nodes when the
  label config hash changes
  * cleanup nodes who's config hash is no longer available in any
    provider
  * is there a way to keep using those nodes? Exclude some config fields
    (e.g. description) from the label config hash as it's not impacting
    the instance.

Change-Id: I14d61719da6d4719cd147edeb6b35d7cae4ee93e