Add pagination to Github graphql canmerge query
This adds pagination support to the Github graphql canmerge query.

There are three items that could theoretically be paginated:

* The list of branch protection rules
* The list of check suites for the commit
* The list of check runs for each suite

We can get 100 of each of these in one query, which works most of the
time, but if these values are exceeded, we may have incomplete
information and may not be able to determine if a change is mergeable.

To account for that, the canmerge query is updated to support
pagination on these three items.  Like the paginated branch protection
rule query, this uses Github global object ids to perform simpler
queries for the second and later pages.

The fake graphql server has been extended to support the additional
interfaces and connections.

One notable quirk: even though everything in the API suggests that it
should be possible to have more than one check run with the same name,
Github itself appears to discard older check runs with the same name
and only returns a single check run (the latest) for a given name.
The existing test infrastructure reflected that, so comments
clarifying the assumption have been added.

The "legacy" canmerge query corresponded to Github Enterprise < 2.21.
That was deprecated on 2021-06-09 and is no longer supported.  To
simplify the implementation, support for the legacy canmerge query is
removed.

Change-Id: I997ebd731bbb502c636580b7e13b4f2940f1be6d