Support github required conversation resolution
As part of its branch protection rules, github has the option to
require that all conversations be "resolved".  Resolving is an
explicit user action and doing so sets a flag on the review thread.

("Conversation" in the UI seems to be a synonym for "review thread"
in the API.)

Unfortunately, whether all conversations are resolved is not exposed
on the pull request as a boolean (like "reviewDecision").  So to
determine if this is the case, we will fetch the review threads as
part of the canMerge query.  In the unlikely event there are more
than 100 threads, we will use pagination to retrieve the remainder
in subsequent queries.  Though as soon as we find the first
unresolved conversation, we will stop.

(It may be possible to infer this state, perhaps along with other
boolean blockers, by examining the mergeStateStatus on the PR.
However, this would likely require some siginificant real-world
data collection to determine if it is sufficiently reliable for
the task.  That is left to a future change.)

Change-Id: I1b4907ad1837548a9ec65a5d5e15b06f57fcdd45