WIP: Add vars-files job attribute
This adds a job attribute which may be used to instruct Zuul to load
variables from a file in a target git repo before running a job's
playbook.

The intended use case is jobs which are heavily reliant on variables
for their configuration and where these veriables change quite
frequently.  When users change job variables, a tenant reconfiguration
is required in order to apply the new values.  In practice, the
new values do not typically change Zuul's behavior, only the input
data supplied to the job.

To make this case more efficient, this allows users to "offload" Zuul
variables into files that are in the repo itself.  Changes to these
dedicated variable files will not trigger Zuul reconfiguration, but
they will achieve the same effect when running jobs.

An obvious question is "why not add an include_vars to playbooks?".
This is a reasonable solution, but it does require some extra work
to anticipate the situations where doing so would be useful.
Additionally, because Zuul runs multiple playbooks, in order to
achieve full parity with Zuul job variables, the import would need
to occur very early in the job (perhaps a base-job pre-run playbook)
and use cacheable facts.  This suggests that the base job itself
might need to provide a Zuul job-variable-based API to load variable
files.  At this point, it is no longer a simple solution, nor is it
as user-friendly as vars-files.

As mentioned above, the primary use case is offloading what would
otherwise be Zuul job variables to a file.  This change attempts
to encode a behavior that is most similar to what happens with
job.vars.  That is to say, if someone moves an existing "job.vars"
variable to a "job.vars-files" file, they should ideally get the
same behavior.

To that end, vars-files assumes (unless otherwise specified) that
the file it should read is in the same repo where the job containing
the vars-files entry is defined.

Users can achieve more advanced effects by specifying that the
file should be read from a different project.

Finally, users may also use this feature in centrally-defined jobs
that may apply to any repo by specifying that the file should be
read from the project under test.  As an example, this would allow
patterns where data such as python or nodejs versions may be defined
in a file in the repo instead of as a project variable.

TODO:
  * tests
  * release note

Change-Id: I7afb0f515e87b884f1e548b3bcf798884e3714e1