Handle kubectl port-foward terminations
The kubernetes project made an interesting and not universally
loved decision[1] to fix an issue where kubectl port-forward could
get stuck in an infinite loop on some errors by causing it to exit
on any error, even benign ones such as a connection refused due
to the target service not yet listening.

This means that, in Zuul, if a playbook ever runs a command task
in a job on a pod before the zuul_console task, the port-forward
will crash and never recover (and so will not be available for
subsequent tasks).  "Don't do that" is one solution to this, but
isn't particularly robust, and there could be other temporary issues
that might similarly cause a disconnect.

Further, it's conceivable that if another job starts a port-foward
that happens to land on the same port, the first job, if it's still
running, might end up streaming the logs from the second job.

To avoid all of this, this change does the following:

1) It runs the port-forward in a shell with an infinite while loop.
   This essentially lets us outsource process monitoring to a shell
   subprocess rather than needing to dedicate an internal thread
   to restarting the port-forward on failure.
2) In order to kill the port-forward (which is now protected from
   signals by the shell), we start it in a process group and kill
   the whole group (as we do for Ansible).
3) Because the port used for each "host" (ie, pod) must be known
   before we start running Ansible, we can no longer let kubectl
   pick the ephemeral local port.  Instead, we will bind a socket
   ourselves to reserve an ephemeral port number, and then provide
   that port number to kubectl.  As long as we aren't listening
   on the socket, kubectl is able to create its own socket and
   bind it.
4) Because of #3, we take extra care to close the socket when
   finished.

A test was written in order to verify the problem and solution
described above, but running it in the gate is currently impractical
since we have no real pod running.  It is likely that we can add
such a system, but that is left to future work.

In the interim though, this change also adds all of the infrastructure
that would be needed to run such a test, including new test-only
environment variables to indicate how to connect to a pod (this
mirrors the way we specify how to connect to a vm).

[1] https://github.com/kubernetes/kubernetes/pull/103526

Change-Id: Iaa5ab61d6dabd08ad9f9e04d78b4d2110ce24c13