Handle ref_protected event attribute from GitLab
When GitLab emits push events, it includes a "ref_protected"
attribute to indicate if the target branch is protected.

This differs from Github which does not include that information
on push events.

The fact that GitLab includes this flag is important because
there may be a race condition which would prevent GitLab from
returning the branch in a query.  This could lead to Zuul receiving
a push event for a branch creation, clearing the branch cache for
that project, querying GitLab for the protection status of the
branch, receiving a 404, then assuming the branch is protected
since that is the Zuul default.

This change implements the following:

* It uses the ref_protected attribute to update the branch cache
  without querying GitLab.
* It no longer clears the branch cache on push events since
  it can receive incremental updates.
* It adds a test to GitLab to assert that a newly-created unprotected
  branch does not cause a reconfiguration.
* It adds the same test to Github for parity.

Change-Id: I08ed62a1555e664d24f88182458517e6d4ed269e