Replace blockdiag graphs with graphviz graphs
Blockdiag and its sphinx extention are no longer maintained. This was
fine until we started to run jobs with python3.12 and blockdiag needed
older Pillow versions to work. And those older Pillow versions need
libjpeg-dev to build against on python3.12.

We switch to using graphviz because it is maintained software, sphinx
has built in support for it, and Zuul is already using it.

To get better margins around the graphviz images we load a custom.css
file with extra padding at the bottom of image divs. Note there was a
preexisting custom.css with css rules in it, but this file was not
loaded before so I have removed all the preexisting rules to preserve
current behavior of those elements.

Change-Id: Ibb63cb4b810d0e1eda7d4fe25cdf40026ed14a95