Replace blockdiag graphs with graphviz graphs
Blockdiag and its sphinx extention are no longer maintained. This was
fine until we started to run jobs with python3.12 and blockdiag needed
older Pillow versions to work. And those older Pillow versions need
libjpeg-dev to build against on python3.12.

We switch to using graphviz because it is maintained software, sphinx
has built in support for it, and Zuul is already using it.

Change-Id: Ibb63cb4b810d0e1eda7d4fe25cdf40026ed14a95