Fix area covered by dependencies cluster in graphviz graph
In the multiple changes with dependencies graphviz graph I modified
things to stop showing a dependency relationship between repo B and repo
C. When I did that the grey background for the dependencies cluster
ended up covering the two status bubbles on the left side of the graph.

We fix that by dropping the same rank specification to get things
aligned and instead recreate the problematic edge between repo B and
repo C but instead of drawing it we mark it invisible. This ensures
everything is in the correct location and the cluster doesn't cover too
much content, but we also don't draw the edge so don't convey dependency
relationships that aren't there.

Change-Id: I9c09cea100f2ac5db98ad267ef1e2da2ecbda557