Remove image permissions check on config loading
We only allow an image build job to be attached to a pipeline in
the project where it is defined.  We enforce this at job freeze
time (since at that point we know exactly what job we're trying
to run).  But we also try to detect this condition as a configuration
error so that we can alert users before things get to that point.

That check is both expensive and flawed.  In production on opendev
we observed what seemed like endless looping and recursion.  This
may be due to a network of job variants which have different
inheritance paths such that if we try to follow all the paths,
we loop.  This doesn't happen when freezing since only certain
job variants are chosen.

Until we fix this, remove the configuration-time check and leave
only the job freeze check.

We may be able to restore this by reducing the scope of what we
check at configuration time.  That will be investigated in a later
change.

Change-Id: I2f173035ff755c4eddc5ec4df999fa540ce98dbf