Finish removing test_image_permissions
A previous commit disabled the test_image_permissions check because
in opendev, we saw a complex job hierarchy produce (possibly infinite)
looping in the image permissions check that is run on configuration
loading.

It doesn't seem likely that we will be able to re-enable this check.

* The only way to know the exact configuration of a job that will be
  used on a pipeline is to actually freeze the job in the context of
  a queue item.  The permissions check that we perform at this time
  (along with the associated test case) remain.
* To guess whether we will have a problem or not requires that we
  examine every possible combination of job variants and the variants
  of their parents recursively.  This can introduce infinite loops.
* If we try to abbreviate this by picking an exemplar variant, we
  could end up picking a configuration that won't actually run, which
  could result in a false negative (which is fine) or a false positive
  (which could permanently prevent someone from changing a
  configuration due to a (false) error that can not be fixed).

It may be possible to detect the case where *all* versions of a job
would trigger a permissions error, and only in the case where there are
no loops.  But this would require significant complexity and
optimization. Given that image build jobs are likely to be created only
by advanced users, and we also want to keep reconfiguration times short,
it is probably better to rely on documentation and the freeze-job check
rather than expend undue effort attempting to catch this error early.

If we find this is a problem in the future, we could revisit this.

Change-Id: Icda38f4d59ace4a25dec0ed2db3dcd449852aac7