Fix race in blobstore put
Give the following circumstances:

* A blobstore entry (e.g., for an item's repo state) associated
  with an item that was recently (within an hour) removed from a pipeline

* An identical entry (same content and key) that is being added
  (perhaps due to a recheck, or other situation that results in the
  same repo state).

If a blob store cleanup happens while the second event is in the
process of putting its data to the blob store, the following
race condition can occur (annotated with scheduler numbers):

[1]: start blob store cleanup
[1]: get start time
[1]: lock pipeline foo.check
[1]: get all live blobs for pipeline foo.check (does not include our blob)
[1]: unlock pipeline foo.check
[2]: get pipeline lock for foo.check
[2]: call put blob for key
[1]: lock key (possibility 1)
[2]: check flag path exists
[1]: lock key (possibility 2)
[1]: recursive delete key (data and flag)
[2]: set flag for key (fails with exception)

This is because when we put a blob in the store, we perform a "check"
on the key before doing so.  If the key is already in the store, we
update its timestamp and return without locking the key.

The cleanup method locks the key, checks the timestamp, and deletes
the key if the timestamp is old.  Our race condition happens when the
put check that we perform outside the lock runs while the delete
method is in progress.

We might consider adding a grace period, but that won't work because
of the way we use timestamps when deleting.  There is no way to use a
grace period to resolve the critical section between checking the
timestamp and deleting during the delete call.

To fix this, we must remove the check from the put call that is
outside the lock.  The one inside the lock will remain and, since it
is inside the lock, will cooperate with the delete call.

In the case where a duplicate value is being inserted, this will add
some extra ZK calls to perform the lock (though we still won't need to
actually insert the data, so it's still less expensive than a typical
put).  But in the case where a new value is being inserted, we will
save the cost of the redundant exists and set calls from the check.

Change-Id: I91b5dc7d477015138fb7daeaa3c164fa7af99261