Fix race in test_state_machines_*
These tests shut down the launcher main thread, but if the test runs
fast enough, they may shut down the launcher thread before it loads
configuration, which means that waitUntilSettled (which verifies that
the launcher and the scheduler are in sync) will never return.

Fix this by waiting until the launcher is settled, then shut down
the main thread and wait for it to finish before starting the test.

Change-Id: Ic516cdbcb9fd11a78a3573928be606b4aede19ad