Add "show all jobs" to the status pipeline overview page
This adds the same "show all jobs" toggle that is on the pipeline
detail page to the overview page.  When enabled, it will cause all
the queue items in all the queues to show their jobs.

If both toggles are enabled, the result is something like the old
status page.

Change-Id: If60f32c28fba79abc6a9b5f6a76f6c5c6bb01ee1