Add "expand all" to the status pipeline overview page
This adds an "expand all" toggle to th overview page.  When enabled,
it will cause all the queue items in all the queues to be expanded and
to show  their jobs.

If both the pipeline toggle and this one are enabled, the result is
something like the old status page.

This is separate from the "show all jobs" toggle on the pipeline summary
page to allow users to have different preferences for each.

Change-Id: If60f32c28fba79abc6a9b5f6a76f6c5c6bb01ee1