Fix handling large command output with split streams
The custom command module is suceptible to deadlocks if used with
split stdout/stderr streams and large volumes of buffered output.  If
a command outputs more than the pipe buffer size (64Kib) in a single
line on both stdout and stderr, our custom command module will attempt
to read the full line from stdout before beginning the read from
stderr.  If the writer is buffering output, then it may be stuck
waiting to finish writing stderr before it will be permitted to
context switch back to filling the stdout pipe.  Since the newline is
never sent on stdout, this results in a deadlock, and the Ansible task
never completes.

  writer: write 70k to internal stdout buffer
  writer: write 70k to internal stderr buffer
  writer: write 64k to stdout pipe
  reader: read 64k from stdout pipe; continue attempting read from stdout
  writer: write 64k to stderr pipe

To correct this, we will spawn a thread each for stdout and stderr so
that they may be service independently.

The streaming test is updated to exercise this; the only result we're
looking for in the test is that the job does not time out.

Change-Id: I0cda3b7c11ce7ae71f8dc915fb36063386d0b8b8