Don't join streaming threads when skipped before connect
We start a log stream thread for every "command/shell" task, and if
Ansible tells us that the task ends up being skipped, we shut down
the streaming thread quickly.  We then wait for the thread to join;
mostly to keep things tidy so that if the thread emits any messages
they happen immediately and not later when we start a new task.

However, if we are still attempting to connect to the host when
the "skipped" callback is invoked, we may spend a few extra seconds
waiting for the connection to fail before joining the thread.

This can happen if a skipped command task is run on a host that is
unreachable or not running a log streaming server.

To speed up this case, let's avoid joining the streaming thread if
we haven't actually completed an initial connection when we stop
the streamer.  Likewise, if we are stopped during connection, let's
return quickly and without further log entries.

Change-Id: I7b039a84a939ad4dc9730661241fe40358d96cfa