Copy provider-related config objects to dynamic layouts
We do not intend to support dynamic (speculative) changes to
provider-related config objects.  When we create a new dynamic
config, we create the new objects and add them to a temporary
layout for validation, then discard that layout.  The real layout
that we return omits these objects entirely.

That means that any dynamic layouts generated for changes in
untrusted projects are running with no provider-related objects at
all.  This effectively means that untrusted projects can't have
functioning provider configs.

To fix this, copy the current "live" provider-related objects from
the active layout to the dynamic layout before returning it.  This
is what we do with pipelines and semaphores -- other objects that
should behave similarly.

Change-Id: Ie1eb2c8c4cde836455fbd713de0e939b7f0cd8a9